fix: cli variables for appCfg
On-behalf-of: @SAP a.shcherbatiuk@sap.com
